chore(Dockerfile): refine Docker multi-stage build
Refine Docker multi-stage build in the Dockerfile to improve security and optimize the container image size. Changed from using golang:1.21-alpine as the base image to using it as a builder and copying the final artifact into a smaller alpine image. This change reduces the attack surface of the runtime container and minimizes its footprint.