fix(svggen): Corrected month rendering in SVG calendar
Corrected an issue in `HandleCalendar` function where the SVG template assumed that all months have exactly 5 weeks. This led to incorrect rendering of months that spread over 6 weeks.

The SVG’s height is now dynamic and will accommodate 6 rows of weeks if necessary, based on the start day and the number of days in the month. Additionally, to handle SVG positioning, the range and rect element values within the SVG code got modified to cater to instances where a 6th week requires a display.

Closes #4