feat(svggen): Emphasise active section of SVG Progress Gauge
This commit aims at refactoring the code responsible for creating the SVG gauge pie sections.
The update involves calculating the active pie section first, and then creating each section accordingly.

The primary change is in how the 'createPiePath' function is implemented. Now, it also accepts an 'isActive' argument that indicates whether the pie section is active or not.
If active, the function increases the inner radius by 10% to increase the representation of the active section, creating a more pronounced visual impact.